requirements.txt
tensorflow>=2.15.0
numpy
matplotlib
seaborn
scikit-learn
httpx
pillow